feat(sockets): added waiting room for students

+++ src/live/rooms.ts
@@ -1,4 +1,9 @@
-import { roomInsertSchema, roomSelectSchema, type RoomInsert } from '$lib/schemas/room.schema';
+import {
+	roomInsertSchema,
+	roomSelectSchema,
+	type Room,
+	type RoomInsert
+} from '$lib/schemas/room.schema';
 import { db } from '$lib/server/db';
 import { roomTable } from '$lib/server/db/schema';
 import { TOPICS } from '$lib/server/topics';
@@ -7,19 +12,37 @@ import type { User } from '$lib/types/user.type';
 import { eq } from 'drizzle-orm/sql/expressions/conditions';
 import { live, LiveError, type LiveContext } from 'svelte-realtime';
 
+async function getRooms(): Promise<Room[]> {
+	const rooms = await db.query.roomTable.findMany({
+		with: {
+			quiz: {
+				columns: {
+					title: true
+				}
+			}
+		}
+	});
+	return roomSelectSchema.array().parse(rooms);
+}
+
+async function getRoomById(roomId: string): Promise<Room | undefined> {
+	const room = await db.query.roomTable.findFirst({
+		with: {
+			quiz: {
+				columns: {
+					title: true
+				}
+			}
+		},
+		where: (room, { eq }) => eq(room.id, roomId)
+	});
+	return room ? roomSelectSchema.parse(room) : undefined;
+}
+
 export const rooms = live.stream(
 	TOPICS.rooms,
 	async () => {
-		const rooms = await db.query.roomTable.findMany({
-			with: {
-				quiz: {
-					columns: {
-						title: true
-					}
-				}
-			}
-		});
-		return roomSelectSchema.array().parse(rooms);
+		return getRooms();
 	},
 	{ merge: 'crud', key: 'id' }
 );
@@ -29,7 +52,16 @@ export const insertRoom = live.validated(
 	async (ctx: LiveContext<User>, roomInsert: RoomInsert) => {
 		if (!ctx.user) throw new LiveError('UNAUTHORIZED', 'User not authenticated');
 		try {
-			const [room] = await db.insert(roomTable).values(roomInsert).returning();
+			const [{ id }] = await db
+				.insert(roomTable)
+				.values(roomInsert)
+				.returning({ id: roomTable.id });
+
+			const room = await getRoomById(id);
+			if (!room) {
+				throw new LiveError('NOT_FOUND', 'Room not found');
+			}
+
 			ctx.publish(TOPICS.rooms, 'created', room);
 			return true;
 		} catch (error) {
@@ -42,9 +74,15 @@ export const insertRoom = live.validated(
 export const deleteRoom = live(async (ctx: LiveContext<User>, roomId: string) => {
 	if (!ctx.user) throw new LiveError('UNAUTHORIZED', 'User not authenticated');
 	try {
+		const room = await getRoomById(roomId);
+		if (!room) {
+			throw new LiveError('NOT_FOUND', 'Room not found');
+		}
+		room.state = 'ended';
+
 		await db.delete(roomTable).where(eq(roomTable.id, roomId)).returning();
 		ctx.publish(TOPICS.rooms, 'deleted', { id: roomId });
-		ctx.publish(TOPICS.room(roomId), 'set', { state: 'ended' });
+		ctx.publish(TOPICS.room(roomId), 'set', room);
 
 		return true;
 	} catch (error) {
@@ -53,21 +91,34 @@ export const deleteRoom = live(async (ctx: LiveContext<User>, roomId: string) =>
 	}
 });
 
+export const startRoom = live(async (ctx: LiveContext<User>, roomId: string) => {
+	if (!ctx.user) throw new LiveError('UNAUTHORIZED', 'User not authenticated');
+
+	try {
+		const room = await getRoomById(roomId);
+		if (!room) {
+			throw new LiveError('NOT_FOUND', 'Room not found');
+		}
+		room.state = 'started';
+
+		await db
+			.update(roomTable)
+			.set({ state: 'started' })
+			.where(eq(roomTable.id, roomId))
+			.returning();
+		ctx.publish(TOPICS.room(roomId), 'set', room);
+		ctx.publish(TOPICS.rooms, 'updated', room);
+	} catch (error) {
+		console.error('Fehler beim Starten des Raums:', error);
+		throw new LiveError('DB', 'Error occurred while starting room');
+	}
+});
+
 export const room = live.room({
 	topic: (_, roomId: string) => TOPICS.room(roomId),
 	merge: 'set',
 	init: async (_, roomId: string) => {
-		const room = await db.query.roomTable.findFirst({
-			with: {
-				quiz: {
-					columns: {
-						title: true
-					}
-				}
-			},
-			where: (room, { eq }) => eq(room.id, roomId)
-		});
-
+		const room = await getRoomById(roomId);
 		return room ?? { state: 'ended' };
 	},
 	presence: (ctx: LiveContext<User>): PresenceUser => ({
